cut3d tests
